Refactor overall JS code (#15)
+ Removed unneeded files.
+ Further code refactoring for seperating routes.
+ Adjusted .gitignore to include future slush files.
