+++ .gitignore
@@ -5,3 +5,4 @@ node_modules
 users.json
 .nvmrc
 .env.sample
+slush/*
