+++ src/app/util/remind.js
@@ -1,0 +1,4 @@
+// run this file periodically to find users who have not accepted the ToS
+const reminders = require('./../routes/interactive/reminder');
+
+reminders.remind();
