--- src/remind.js
@@ -1,4 +1,0 @@
-// run this file periodically to find users who have not accepted the ToS
-const onboard = require('./onboard');
-
-onboard.remind();
